pyproject.toml
[project]
name = "spirsa"
version = "0.1.0"
description = "Art portfolio for Štefica Pirša"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "gunicorn>=23.0",
    "pillow>=11.0",
    "psycopg[binary]>=3.2",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11",
]

[tool.ruff]
line-length = 100
exclude = ["*/migrations/", "manage.py"]

[tool.ruff.lint]
extend-select = ["I", "UP"]
